Refactor: Functions and registerScreen
